Some quick cleanup on the graphql api
